addBeer function working + fancy

+++ app/src/main/java/com/example/alcoholcounter/database/Drink.java
@@ -20,7 +20,7 @@ public class Drink {
     public double unit;
 
     @ColumnInfo(name = "quantity")
-    public double quantity;
+    public int quantity;
 
     @ColumnInfo(name = "degrees")
     public double degrees;
